requirements.txt
flask
flask-sqlalchemy
mysqlclient
MySQLdb
python-dotenv